# Shubhankar/totorial/server/public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"Company": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"employees": {
						"items": {
							"$ref": "#/components/schemas/Employee"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"address",
					"email",
					"employees"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Employee": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"companyid": {
						"type": "number",
						"format": "double"
					},
					"company": {
						"$ref": "#/components/schemas/Company"
					}
				},
				"required": [
					"id",
					"firstName",
					"lastName",
					"content",
					"companyid",
					"company"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICompanyData": {
				"properties": {
					"Name": {
						"type": "string"
					},
					"Address": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"Name",
					"Address",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Games": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"players": {
						"items": {
							"$ref": "#/components/schemas/Player"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"players"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Player": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"games": {
						"items": {
							"$ref": "#/components/schemas/Games"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"address",
					"games"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPlayerPayload": {
				"properties": {
					"title": {
						"type": "string"
					},
					"tutorialId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"title",
					"tutorialId"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "express-typescript",
		"version": "1.1.1",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/Companys": {
			"get": {
				"operationId": "GetCompany",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Company"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Company"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateCompany",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Company"
								}
							}
						}
					}
				},
				"tags": [
					"Company"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICompanyData"
							}
						}
					}
				}
			}
		},
		"/Companys/{id}": {
			"get": {
				"operationId": "GetCompanys",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Company"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Company"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteCompanys",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Company"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Company"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/players": {
			"post": {
				"operationId": "CreatePlayer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Player"
								}
							}
						}
					}
				},
				"tags": [
					"players"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPlayerPayload"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetPlayers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Player"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"players"
				],
				"security": [],
				"parameters": []
			}
		},
		"/players/{id}": {
			"get": {
				"operationId": "GetPlayer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Player"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"players"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}